Semantic mistake referring to request not response
The response for the AdminCreateUser flow **has to** include both the `event.request.usernameParameter` and `event.request.codeParameter`. If this is not the case, the default message is used instead and no warning is given to the user. The top line of the documentation suggests that the `event.request.usernameParameter` is the required **request** parameter, and not **response** parameter.